Created packages for each minigame
