--- src/main/java/io/github/hypixel_api_wrapper/wrapper/games/HypixelBattleground.java
@@ -1,18 +1,0 @@
-package io.github.hypixel_api_wrapper.wrapper.games;
-
-public class HypixelBattleground implements HypixelGame {
-    @Override
-    public String getGameID() {
-        return "BATTLEGROUND";
-    }
-
-    @Override
-    public String getGameName() {
-        return "Warlords";
-    }
-
-    @Override
-    public boolean isNetworkPresent() {
-        return true;
-    }
-}
