--- src/main/java/io/github/hypixel_api_wrapper/wrapper/games/HypixelBuildBattle.java
@@ -1,18 +1,0 @@
-package io.github.hypixel_api_wrapper.wrapper.games;
-
-public class HypixelBuildBattle implements HypixelGame {
-    @Override
-    public String getGameID() {
-        return "BUILD_BATTLE";
-    }
-
-    @Override
-    public String getGameName() {
-        return "Build Battle";
-    }
-
-    @Override
-    public boolean isNetworkPresent() {
-        return true;
-    }
-}
