--- src/main/java/io/github/hypixel_api_wrapper/wrapper/games/HypixelCrazyWalls.java
@@ -1,18 +1,0 @@
-package io.github.hypixel_api_wrapper.wrapper.games;
-
-public class HypixelCrazyWalls implements HypixelGame {
-    @Override
-    public String getGameID() {
-        return "TRUE_COMBAT";
-    }
-
-    @Override
-    public String getGameName() {
-        return "TrueCombat";
-    }
-
-    @Override
-    public boolean isNetworkPresent() {
-        return false;
-    }
-}
