--- src/main/java/io/github/hypixel_api_wrapper/wrapper/games/HypixelDuels.java
@@ -1,18 +1,0 @@
-package io.github.hypixel_api_wrapper.wrapper.games;
-
-public class HypixelDuels implements HypixelGame {
-    @Override
-    public String getGameID() {
-        return "DUELS";
-    }
-
-    @Override
-    public String getGameName() {
-        return "Duels";
-    }
-
-    @Override
-    public boolean isNetworkPresent() {
-        return true;
-    }
-}
