--- src/main/java/io/github/hypixel_api_wrapper/wrapper/games/HypixelHousing.java
@@ -1,18 +1,0 @@
-package io.github.hypixel_api_wrapper.wrapper.games;
-
-public class HypixelHousing implements HypixelGame {
-    @Override
-    public String getGameID() {
-        return "HOUSING";
-    }
-
-    @Override
-    public String getGameName() {
-        return "Housing";
-    }
-
-    @Override
-    public boolean isNetworkPresent() {
-        return true;
-    }
-}
