--- src/main/java/io/github/hypixel_api_wrapper/wrapper/games/HypixelHungerGames.java
@@ -1,18 +1,0 @@
-package io.github.hypixel_api_wrapper.wrapper.games;
-
-public class HypixelHungerGames implements HypixelGame {
-    @Override
-    public String getGameID() {
-        return "SURVIVAL_GAMES";
-    }
-
-    @Override
-    public String getGameName() {
-        return "Blitz Survival Games";
-    }
-
-    @Override
-    public boolean isNetworkPresent() {
-        return true;
-    }
-}
