--- src/main/java/io/github/hypixel_api_wrapper/wrapper/games/HypixelPaintball.java
@@ -1,18 +1,0 @@
-package io.github.hypixel_api_wrapper.wrapper.games;
-
-public class HypixelPaintball implements HypixelGame {
-    @Override
-    public String getGameID() {
-        return "PAINTBALL";
-    }
-
-    @Override
-    public String getGameName() {
-        return "Paintball";
-    }
-
-    @Override
-    public boolean isNetworkPresent() {
-        return true;
-    }
-}
