--- src/main/java/io/github/hypixel_api_wrapper/wrapper/games/HypixelPit.java
@@ -1,18 +1,0 @@
-package io.github.hypixel_api_wrapper.wrapper.games;
-
-public class HypixelPit implements HypixelGame {
-    @Override
-    public String getGameID() {
-        return "PIT";
-    }
-
-    @Override
-    public String getGameName() {
-        return "Pit";
-    }
-
-    @Override
-    public boolean isNetworkPresent() {
-        return true;
-    }
-}
