--- src/main/java/io/github/hypixel_api_wrapper/wrapper/games/HypixelQuake.java
@@ -1,18 +1,0 @@
-package io.github.hypixel_api_wrapper.wrapper.games;
-
-public class HypixelQuake implements HypixelGame {
-    @Override
-    public String getGameID() {
-        return "QUAKECRAFT";
-    }
-
-    @Override
-    public String getGameName() {
-        return "Quake";
-    }
-
-    @Override
-    public boolean isNetworkPresent() {
-        return true;
-    }
-}
