--- src/main/java/io/github/hypixel_api_wrapper/wrapper/games/HypixelSMP.java
@@ -1,18 +1,0 @@
-package io.github.hypixel_api_wrapper.wrapper.games;
-
-public class HypixelSMP implements HypixelGame {
-    @Override
-    public String getGameID() {
-        return "SMP";
-    }
-
-    @Override
-    public String getGameName() {
-        return "SMP";
-    }
-
-    @Override
-    public boolean isNetworkPresent() {
-        return true;
-    }
-}
