--- src/main/java/io/github/hypixel_api_wrapper/wrapper/games/HypixelSkyClash.java
@@ -1,18 +1,0 @@
-package io.github.hypixel_api_wrapper.wrapper.games;
-
-public class HypixelSkyClash implements HypixelGame {
-    @Override
-    public String getGameID() {
-        return "SKYCLASH";
-    }
-
-    @Override
-    public String getGameName() {
-        return "SkyClash";
-    }
-
-    @Override
-    public boolean isNetworkPresent() {
-        return false;
-    }
-}
