--- src/main/java/io/github/hypixel_api_wrapper/wrapper/games/HypixelSkyWars.java
@@ -1,18 +1,0 @@
-package io.github.hypixel_api_wrapper.wrapper.games;
-
-public class HypixelSkyWars implements HypixelGame {
-    @Override
-    public String getGameID() {
-        return "SKYWARS";
-    }
-
-    @Override
-    public String getGameName() {
-        return "SkyWars";
-    }
-
-    @Override
-    public boolean isNetworkPresent() {
-        return true;
-    }
-}
