--- src/main/java/io/github/hypixel_api_wrapper/wrapper/games/HypixelSkyblock.java
@@ -1,18 +1,0 @@
-package io.github.hypixel_api_wrapper.wrapper.games;
-
-public class HypixelSkyblock implements HypixelGame {
-    @Override
-    public String getGameID() {
-        return "SKYBLOCK";
-    }
-
-    @Override
-    public String getGameName() {
-        return "SkyBlock";
-    }
-
-    @Override
-    public boolean isNetworkPresent() {
-        return true;
-    }
-}
