--- src/main/java/io/github/hypixel_api_wrapper/wrapper/games/HypixelSuperSmash.java
@@ -1,18 +1,0 @@
-package io.github.hypixel_api_wrapper.wrapper.games;
-
-public class HypixelSuperSmash implements HypixelGame {
-    @Override
-    public String getGameID() {
-        return "SUPER_SMASH";
-    }
-
-    @Override
-    public String getGameName() {
-        return "Smash Heros";
-    }
-
-    @Override
-    public boolean isNetworkPresent() {
-        return true;
-    }
-}
