--- src/main/java/io/github/hypixel_api_wrapper/wrapper/games/HypixelTNTGames.java
@@ -1,18 +1,0 @@
-package io.github.hypixel_api_wrapper.wrapper.games;
-
-public class HypixelTNTGames implements HypixelGame {
-    @Override
-    public String getGameID() {
-        return "TNTGAMES";
-    }
-
-    @Override
-    public String getGameName() {
-        return "TNT Games";
-    }
-
-    @Override
-    public boolean isNetworkPresent() {
-        return true;
-    }
-}
