--- src/main/java/io/github/hypixel_api_wrapper/wrapper/games/HypixelTurboKart.java
@@ -1,18 +1,0 @@
-package io.github.hypixel_api_wrapper.wrapper.games;
-
-public class HypixelTurboKart implements HypixelGame {
-    @Override
-    public String getGameID() {
-        return "GINGERBREAD";
-    }
-
-    @Override
-    public String getGameName() {
-        return "Turbo Kart Racers";
-    }
-
-    @Override
-    public boolean isNetworkPresent() {
-        return true;
-    }
-}
