--- src/main/java/io/github/hypixel_api_wrapper/wrapper/games/HypixelUHC.java
@@ -1,18 +1,0 @@
-package io.github.hypixel_api_wrapper.wrapper.games;
-
-public class HypixelUHC implements HypixelGame {
-    @Override
-    public String getGameID() {
-        return "UHC";
-    }
-
-    @Override
-    public String getGameName() {
-        return "UHC Champions";
-    }
-
-    @Override
-    public boolean isNetworkPresent() {
-        return true;
-    }
-}
