--- src/main/java/io/github/hypixel_api_wrapper/wrapper/games/HypixelVampireZ.java
@@ -1,18 +1,0 @@
-package io.github.hypixel_api_wrapper.wrapper.games;
-
-public class HypixelVampireZ implements HypixelGame {
-    @Override
-    public String getGameID() {
-        return "VAMPIREZ";
-    }
-
-    @Override
-    public String getGameName() {
-        return "VampireZ";
-    }
-
-    @Override
-    public boolean isNetworkPresent() {
-        return true;
-    }
-}
