--- src/main/java/io/github/hypixel_api_wrapper/wrapper/games/HypixelWalls.java
@@ -1,18 +1,0 @@
-package io.github.hypixel_api_wrapper.wrapper.games;
-
-public class HypixelWalls implements HypixelGame {
-    @Override
-    public String getGameID() {
-        return "WALLS";
-    }
-
-    @Override
-    public String getGameName() {
-        return "Walls";
-    }
-
-    @Override
-    public boolean isNetworkPresent() {
-        return false;
-    }
-}
