--- src/main/java/io/github/hypixel_api_wrapper/wrapper/games/HypixelWalls3.java
@@ -1,18 +1,0 @@
-package io.github.hypixel_api_wrapper.wrapper.games;
-
-public class HypixelWalls3 implements HypixelGame {
-    @Override
-    public String getGameID() {
-        return "WALLS3";
-    }
-
-    @Override
-    public String getGameName() {
-        return "Mega Walls";
-    }
-
-    @Override
-    public boolean isNetworkPresent() {
-        return true;
-    }
-}
