+++ src/main/java/io/github/hypixel_api_wrapper/wrapper/games/bedwars/HypixelBedWars.java
@@ -1,0 +1,20 @@
+package io.github.hypixel_api_wrapper.wrapper.games.bedwars;
+
+import io.github.hypixel_api_wrapper.wrapper.games.HypixelGame;
+
+public class HypixelBedWars implements HypixelGame {
+    @Override
+    public String getGameID() {
+        return "BEDWARS";
+    }
+
+    @Override
+    public String getGameName() {
+        return "Bed Wars";
+    }
+
+    @Override
+    public boolean isNetworkPresent() {
+        return true;
+    }
+}
