+++ src/main/java/io/github/hypixel_api_wrapper/wrapper/games/classic/turbo_kart/HypixelTurboKart.java
@@ -1,0 +1,20 @@
+package io.github.hypixel_api_wrapper.wrapper.games.classic.turbo_kart;
+
+import io.github.hypixel_api_wrapper.wrapper.games.HypixelGame;
+
+public class HypixelTurboKart implements HypixelGame {
+    @Override
+    public String getGameID() {
+        return "GINGERBREAD";
+    }
+
+    @Override
+    public String getGameName() {
+        return "Turbo Kart Racers";
+    }
+
+    @Override
+    public boolean isNetworkPresent() {
+        return true;
+    }
+}
