+++ src/main/java/io/github/hypixel_api_wrapper/wrapper/games/classic/vampireZ/HypixelVampireZ.java
@@ -1,0 +1,20 @@
+package io.github.hypixel_api_wrapper.wrapper.games.classic.vampireZ;
+
+import io.github.hypixel_api_wrapper.wrapper.games.HypixelGame;
+
+public class HypixelVampireZ implements HypixelGame {
+    @Override
+    public String getGameID() {
+        return "VAMPIREZ";
+    }
+
+    @Override
+    public String getGameName() {
+        return "VampireZ";
+    }
+
+    @Override
+    public boolean isNetworkPresent() {
+        return true;
+    }
+}
