+++ src/main/java/io/github/hypixel_api_wrapper/wrapper/games/mcgo/HypixelMCGO.java
@@ -1,0 +1,20 @@
+package io.github.hypixel_api_wrapper.wrapper.games.mcgo;
+
+import io.github.hypixel_api_wrapper.wrapper.games.HypixelGame;
+
+public class HypixelMCGO implements HypixelGame {
+    @Override
+    public String getGameID() {
+        return "MCGO";
+    }
+
+    @Override
+    public String getGameName() {
+        return "Cops and Crims";
+    }
+
+    @Override
+    public boolean isNetworkPresent() {
+        return true;
+    }
+}
