+++ src/main/java/io/github/hypixel_api_wrapper/wrapper/games/skyblock/HypixelSkyblock.java
@@ -1,0 +1,20 @@
+package io.github.hypixel_api_wrapper.wrapper.games.skyblock;
+
+import io.github.hypixel_api_wrapper.wrapper.games.HypixelGame;
+
+public class HypixelSkyblock implements HypixelGame {
+    @Override
+    public String getGameID() {
+        return "SKYBLOCK";
+    }
+
+    @Override
+    public String getGameName() {
+        return "SkyBlock";
+    }
+
+    @Override
+    public boolean isNetworkPresent() {
+        return true;
+    }
+}
