+++ src/main/java/io/github/hypixel_api_wrapper/wrapper/games/super_smash/HypixelSuperSmash.java
@@ -1,0 +1,20 @@
+package io.github.hypixel_api_wrapper.wrapper.games.super_smash;
+
+import io.github.hypixel_api_wrapper.wrapper.games.HypixelGame;
+
+public class HypixelSuperSmash implements HypixelGame {
+    @Override
+    public String getGameID() {
+        return "SUPER_SMASH";
+    }
+
+    @Override
+    public String getGameName() {
+        return "Smash Heros";
+    }
+
+    @Override
+    public boolean isNetworkPresent() {
+        return true;
+    }
+}
