+++ src/main/java/io/github/hypixel_api_wrapper/wrapper/games/tnt_games/HypixelTNTGames.java
@@ -1,0 +1,20 @@
+package io.github.hypixel_api_wrapper.wrapper.games.tnt_games;
+
+import io.github.hypixel_api_wrapper.wrapper.games.HypixelGame;
+
+public class HypixelTNTGames implements HypixelGame {
+    @Override
+    public String getGameID() {
+        return "TNTGAMES";
+    }
+
+    @Override
+    public String getGameName() {
+        return "TNT Games";
+    }
+
+    @Override
+    public boolean isNetworkPresent() {
+        return true;
+    }
+}
